# Optimize distributor push on error