--- a/pkg/util/validation/errors.go
+++ b/pkg/util/validation/errors.go
@@ -2,25 +2,17 @@ package validation
 
 import (
 	"fmt"
-	"net/http"
 	"strings"
 
 	"github.com/prometheus/common/model"
-	"github.com/weaveworks/common/httpgrpc"
 
 	"github.com/cortexproject/cortex/pkg/cortexpb"
 )
 
-// ValidationError is an error returned by series validation. It provides a function to convert
-// the error into an HTTPGRPC error.
+// ValidationError is an error returned by series validation.
 //
 // nolint:golint ignore stutter warning
-type ValidationError interface {
-	error
-
-	// ToHTTPGRPCError returns the httpgrpc version of the error.
-	ToHTTPGRPCError() error
-}
+type ValidationError error
 
 // genericValidationError is a basic implementation of ValidationError which can be used when the
 // error format only contains the cause and the series.
@@ -30,12 +22,8 @@ type genericValidationError struct {
 	series  []cortexpb.LabelAdapter
 }
 
-func (e *genericValidationError) ToHTTPGRPCError() error {
-	return httpgrpc.Errorf(http.StatusBadRequest, e.message, e.cause, formatLabelSet(e.series))
-}
-
 func (e *genericValidationError) Error() string {
-	return e.ToHTTPGRPCError().Error()
+	return fmt.Sprintf(e.message, e.cause, formatLabelSet(e.series))
 }
 
 func newLabelNameTooLongError(series []cortexpb.LabelAdapter, labelName string) ValidationError {
@@ -90,15 +78,10 @@ func newTooManyLabelsError(series []cortexpb.LabelAdapter, limit int) Validation
 	}
 }
 
-func (e *tooManyLabelsError) ToHTTPGRPCError() error {
-	return httpgrpc.Errorf(
-		http.StatusBadRequest,
+func (e *tooManyLabelsError) Error() string {
+	return fmt.Sprintf(
 		"series has too many labels (actual: %d, limit: %d) series: '%s'",
 		len(e.series), e.limit, cortexpb.FromLabelAdaptersToMetric(e.series).String())
-}
-
-func (e *tooManyLabelsError) Error() string {
-	return e.ToHTTPGRPCError().Error()
 }
 
 type noMetricNameError struct{}
@@ -107,12 +90,8 @@ func newNoMetricNameError() ValidationError {
 	return &noMetricNameError{}
 }
 
-func (e *noMetricNameError) ToHTTPGRPCError() error {
-	return httpgrpc.Errorf(http.StatusBadRequest, "sample missing metric name")
-}
-
 func (e *noMetricNameError) Error() string {
-	return e.ToHTTPGRPCError().Error()
+	return "sample missing metric name"
 }
 
 type invalidMetricNameError struct {
@@ -125,12 +104,8 @@ func newInvalidMetricNameError(metricName string) ValidationError {
 	}
 }
 
-func (e *invalidMetricNameError) ToHTTPGRPCError() error {
-	return httpgrpc.Errorf(http.StatusBadRequest, "sample invalid metric name: %.200q", e.metricName)
-}
-
 func (e *invalidMetricNameError) Error() string {
-	return e.ToHTTPGRPCError().Error()
+	return fmt.Sprintf("sample invalid metric name: %.200q", e.metricName)
 }
 
 // sampleValidationError is a ValidationError implementation suitable for sample validation errors.
@@ -140,12 +115,8 @@ type sampleValidationError struct {
 	timestamp  int64
 }
 
-func (e *sampleValidationError) ToHTTPGRPCError() error {
-	return httpgrpc.Errorf(http.StatusBadRequest, e.message, e.timestamp, e.metricName)
-}
-
 func (e *sampleValidationError) Error() string {
-	return e.ToHTTPGRPCError().Error()
+	return fmt.Sprintf(e.message, e.timestamp, e.metricName)
 }
 
 func newSampleTimestampTooOldError(metricName string, timestamp int64) ValidationError {