--- a/pkg/util/validation/validate.go
+++ b/pkg/util/validation/validate.go
@@ -1,7 +1,6 @@
 package validation
 
 import (
-	"fmt"
 	"net/http"
 	"strings"
 	"time"
@@ -30,17 +29,6 @@ const (
 	metricNameTooLong = "metric_name_too_long"
 	helpTooLong       = "help_too_long"
 	unitTooLong       = "unit_too_long"
-
-	errMissingMetricName  = "sample missing metric name"
-	errInvalidMetricName  = "sample invalid metric name: %.200q"
-	errInvalidLabel       = "sample invalid label: %.200q metric %.200q"
-	errLabelNameTooLong   = "label name too long: %.200q metric %.200q"
-	errLabelValueTooLong  = "label value too long: %.200q metric %.200q"
-	errTooManyLabels      = "series has too many labels (actual: %d, limit: %d) series: '%s'"
-	errTooOld             = "timestamp too old: %d metric: %.200q"
-	errTooNew             = "timestamp too new: %d metric: %.200q"
-	errDuplicateLabelName = "duplicate label name: %.200q metric %.200q"
-	errLabelsNotSorted    = "labels not sorted: %.200q metric %.200q"
 
 	// ErrQueryTooLong is used in chunk store, querier and query frontend.
 	ErrQueryTooLong = "the query time range exceeds the limit (query length: %s, limit: %s)"
@@ -94,16 +82,17 @@ type SampleValidationConfig interface {
 	CreationGracePeriod(userID string) time.Duration
 }
 
-// ValidateSample returns an err if the sample is invalid.
-func ValidateSample(cfg SampleValidationConfig, userID string, metricName string, s cortexpb.Sample) error {
+// ValidateSample returns an err if the sample is invalid. The returned ValidationError
+// provides a function to convert it to an HTTPGRPC error.
+func ValidateSample(cfg SampleValidationConfig, userID string, metricName string, s cortexpb.Sample) ValidationError {
 	if cfg.RejectOldSamples(userID) && model.Time(s.TimestampMs) < model.Now().Add(-cfg.RejectOldSamplesMaxAge(userID)) {
 		DiscardedSamples.WithLabelValues(greaterThanMaxSampleAge, userID).Inc()
-		return httpgrpc.Errorf(http.StatusBadRequest, errTooOld, model.Time(s.TimestampMs), metricName)
+		return newSampleTimestampTooOldError(metricName, s.TimestampMs)
 	}
 
 	if model.Time(s.TimestampMs) > model.Now().Add(cfg.CreationGracePeriod(userID)) {
 		DiscardedSamples.WithLabelValues(tooFarInFuture, userID).Inc()
-		return httpgrpc.Errorf(http.StatusBadRequest, errTooNew, model.Time(s.TimestampMs), metricName)
+		return newSampleTimestampTooNewError(metricName, s.TimestampMs)
 	}
 
 	return nil
@@ -117,61 +106,51 @@ type LabelValidationConfig interface {
 	MaxLabelValueLength(userID string) int
 }
 
-// ValidateLabels returns an err if the labels are invalid.
-func ValidateLabels(cfg LabelValidationConfig, userID string, ls []cortexpb.LabelAdapter, skipLabelNameValidation bool) error {
+// ValidateLabels returns an err if the labels are invalid. The returned ValidationError
+// provides a function to convert it to an HTTPGRPC error.
+func ValidateLabels(cfg LabelValidationConfig, userID string, ls []cortexpb.LabelAdapter, skipLabelNameValidation bool) ValidationError {
 	if cfg.EnforceMetricName(userID) {
 		metricName, err := extract.MetricNameFromLabelAdapters(ls)
 		if err != nil {
 			DiscardedSamples.WithLabelValues(missingMetricName, userID).Inc()
-			return httpgrpc.Errorf(http.StatusBadRequest, errMissingMetricName)
+			return newNoMetricNameError()
 		}
 
 		if !model.IsValidMetricName(model.LabelValue(metricName)) {
 			DiscardedSamples.WithLabelValues(invalidMetricName, userID).Inc()
-			return httpgrpc.Errorf(http.StatusBadRequest, errInvalidMetricName, metricName)
+			return newInvalidMetricNameError(metricName)
 		}
 	}
 
 	numLabelNames := len(ls)
 	if numLabelNames > cfg.MaxLabelNamesPerSeries(userID) {
 		DiscardedSamples.WithLabelValues(maxLabelNamesPerSeries, userID).Inc()
-		return httpgrpc.Errorf(http.StatusBadRequest, errTooManyLabels, numLabelNames, cfg.MaxLabelNamesPerSeries(userID), cortexpb.FromLabelAdaptersToMetric(ls).String())
+		return newTooManyLabelsError(ls, cfg.MaxLabelNamesPerSeries(userID))
 	}
 
 	maxLabelNameLength := cfg.MaxLabelNameLength(userID)
 	maxLabelValueLength := cfg.MaxLabelValueLength(userID)
 	lastLabelName := ""
 	for _, l := range ls {
-		var errTemplate string
-		var reason string
-		var cause interface{}
 		if !skipLabelNameValidation && !model.LabelName(l.Name).IsValid() {
-			reason = invalidLabel
-			errTemplate = errInvalidLabel
-			cause = l.Name
+			DiscardedSamples.WithLabelValues(invalidLabel, userID).Inc()
+			return newInvalidLabelError(ls, l.Name)
 		} else if len(l.Name) > maxLabelNameLength {
-			reason = labelNameTooLong
-			errTemplate = errLabelNameTooLong
-			cause = l.Name
+			DiscardedSamples.WithLabelValues(labelNameTooLong, userID).Inc()
+			return newLabelNameTooLongError(ls, l.Name)
 		} else if len(l.Value) > maxLabelValueLength {
-			reason = labelValueTooLong
-			errTemplate = errLabelValueTooLong
-			cause = l.Value
+			DiscardedSamples.WithLabelValues(labelValueTooLong, userID).Inc()
+			return newLabelValueTooLongError(ls, l.Value)
 		} else if cmp := strings.Compare(lastLabelName, l.Name); cmp >= 0 {
 			if cmp == 0 {
-				reason = duplicateLabelNames
-				errTemplate = errDuplicateLabelName
-				cause = l.Name
+				DiscardedSamples.WithLabelValues(duplicateLabelNames, userID).Inc()
+				return newDuplicatedLabelError(ls, l.Name)
 			} else {
-				reason = labelsNotSorted
-				errTemplate = errLabelsNotSorted
-				cause = l.Name
+				DiscardedSamples.WithLabelValues(labelsNotSorted, userID).Inc()
+				return newLabelsNotSortedError(ls, l.Name)
 			}
 		}
-		if errTemplate != "" {
-			DiscardedSamples.WithLabelValues(reason, userID).Inc()
-			return httpgrpc.Errorf(http.StatusBadRequest, errTemplate, cause, formatLabelSet(ls))
-		}
+
 		lastLabelName = l.Name
 	}
 	return nil
@@ -214,32 +193,6 @@ func ValidateMetadata(cfg MetadataValidationConfig, userID string, metadata *cor
 	}
 
 	return nil
-}
-
-// this function formats label adapters as a metric name with labels, while preserving
-// label order, and keeping duplicates. If there are multiple "__name__" labels, only
-// first one is used as metric name, other ones will be included as regular labels.
-func formatLabelSet(ls []cortexpb.LabelAdapter) string {
-	metricName, hasMetricName := "", false
-
-	labelStrings := make([]string, 0, len(ls))
-	for _, l := range ls {
-		if l.Name == model.MetricNameLabel && !hasMetricName && l.Value != "" {
-			metricName = l.Value
-			hasMetricName = true
-		} else {
-			labelStrings = append(labelStrings, fmt.Sprintf("%s=%q", l.Name, l.Value))
-		}
-	}
-
-	if len(labelStrings) == 0 {
-		if hasMetricName {
-			return metricName
-		}
-		return "{}"
-	}
-
-	return fmt.Sprintf("%s{%s}", metricName, strings.Join(labelStrings, ", "))
 }
 
 func DeletePerUserValidationMetrics(userID string, log log.Logger) {